Player and zombie implementation complete
players head is not looking at aim pos

+++ Assets/_Scripts/WeaponManager.cs
@@ -14,7 +14,7 @@ public class WeaponManager : MonoBehaviour
     [SerializeField] Transform barrelPos;
     [SerializeField] float bulletVelocity;
     [SerializeField] int bulletPerShot;
-    public float damage = 20;
+    public int damage = 20;
     AimStateManager aim;
 
     [SerializeField] AudioClip gunShot;
